implement the rest of clang-tidy fixes that isn't m_options
